Move calculate shares function to types package
